A basic game guess game.
The game asks the user to guess a randomly generated number. If the user guesses wrongly, the score gets reduced by 1. 
If the high score is lesser than the score at that very moment, the high score updates to the current score.
It's a basic HTML, CSS & Javascript implementation game.